requirements.txt
absl-py==2.1.0
arandr==0.1.11
async-timeout==4.0.2
attrs==24.2.0
av==10.0.0
certifi==2022.9.24
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorzero==2.0
contourpy==1.3.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
distro==1.8.0
docopt==0.6.2
flatbuffers==20181003210633
fonttools==4.53.1
gpiozero==2.0.1
idna==3.3
ifaddr==0.1.7
jax==0.4.31
jaxlib==0.4.31
kiwisolver==1.4.7
lgpio==0.2.2.0
libevdev==0.5
matplotlib==3.9.2
mediapipe==0.10.15
mido==1.2.10
ml_dtypes==0.5.0
MouseInfo==0.1.3
netifaces==0.11.0
numpy==1.24.2
olefile==0.46
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
patchbox-cli==1.4.0
pexpect==4.8.0
picamera2==0.3.21
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==9.4.0
protobuf==4.25.4
ptyprocess==0.7.0
PyAutoGUI==0.9.54
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
PyGetWindow==0.0.9
PyGObject==3.42.2
pyliblo==0.10.0
PyMsgBox==1.0.9
PyOpenGL==3.1.6
pyparsing==3.1.4
pyperclip==1.9.0
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyRect==0.2.0
PyScreeze==1.0.1
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-osc==1.9.0
python-prctl==1.8.1
python-rtmidi==1.4.7
python3-xlib==0.15
pytweening==1.2.0
pyudev==0.24.0
PyYAML==6.0.2
reportlab==3.6.12
requests==2.28.1
RPi.GPIO==0.7.1a4
scipy==1.14.1
screeninfo==0.8.1
simplejpeg==1.6.6
six==1.16.0
smbus2==0.4.2
sounddevice==0.5.0
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
touchosc2midi==0.0.10
urllib3==1.26.12
urwid==2.1.2
v4l2-python3==0.3.5
zeroconf==0.47.3
